arb-scanner/requirements.txt
websockets==12.0
redis==5.0.1
python-telegram-bot==20.7
stripe==8.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.6.1
aiohttp==3.9.1
python-dotenv==1.0.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
aiosqlite==0.19.0
